Flatten SortField heirarchy and use own file
