+++ src/main/java/jobicade/betterhud/proxy/ClientProxy.java
@@ -8,10 +8,10 @@ import jobicade.betterhud.BetterHud;
 import jobicade.betterhud.config.ConfigManager;
 import jobicade.betterhud.config.HudConfig;
 import jobicade.betterhud.element.HudElement;
-import jobicade.betterhud.element.HudElement.SortType;
 import jobicade.betterhud.gui.GuiHudMenu;
 import jobicade.betterhud.registry.HudElements;
 import jobicade.betterhud.registry.HudRegistryEvent;
+import jobicade.betterhud.registry.SortField;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.IReloadableResourceManager;
 import net.minecraft.client.resources.IResourceManager;
@@ -46,7 +46,7 @@ public class ClientProxy implements HudSidedProxy {
             IReloadableResourceManager reloadableManager = (IReloadableResourceManager)manager;
 
             // Language dictates alphabetical order
-            reloadableManager.registerReloadListener(m -> HudElements.get().invalidateSorts(SortType.ALPHABETICAL));
+            reloadableManager.registerReloadListener(m -> HudElements.get().invalidateSorts(SortField.ALPHABETICAL));
             reloadableManager.registerReloadListener(configManager);
         } else {
             BetterHud.getLogger().warn("Unable to register alphabetical sort update on language change");
