+++ src/main/java/jobicade/betterhud/registry/SortField.java
@@ -1,0 +1,42 @@
+package jobicade.betterhud.registry;
+
+import java.util.Comparator;
+
+import jobicade.betterhud.element.HudElement;
+
+public enum SortField implements Comparator<HudElement<?>> {
+	ALPHABETICAL("alphabetical", false) {
+		@Override
+		public int compare(HudElement<?> a, HudElement<?> b) {
+			return a.getLocalizedName().compareTo(b.getLocalizedName());
+		}
+	}, ENABLED("enabled", false) {
+		@Override
+		public int compare(HudElement<?> a, HudElement<?> b) {
+			int compare = b.settings.get().compareTo(a.settings.get());
+			return compare != 0 ? compare : ALPHABETICAL.compare(a, b);
+		}
+	}, PRIORITY("priority", false) {
+		@Override
+		public int compare(HudElement<?> a, HudElement<?> b) {
+			int compare = a.settings.priority.get().compareTo(b.settings.priority.get());
+			return compare != 0 ? compare : ALPHABETICAL.compare(a, b);
+		}
+	};
+
+	private final String unlocalizedName;
+	private final boolean inverted;
+
+	SortField(String unlocalizedName, boolean inverted) {
+		this.unlocalizedName = "betterHud.menu." + unlocalizedName;
+		this.inverted = inverted;
+	}
+
+	public String getUnlocalizedName() {
+		return unlocalizedName;
+	}
+
+	public boolean isInverted() {
+		return inverted;
+	}
+}
