--- src/main/java/jobicade/betterhud/util/SortField.java
@@ -1,8 +1,0 @@
-package jobicade.betterhud.util;
-
-import java.util.Comparator;
-
-public interface SortField<T> extends Comparator<T> {
-	public String getUnlocalizedName();
-	public boolean isInverted();
-}
